refactor(be): QueryDSL을 사용하는 인터페이스와 구현체 분리
* JPA 관례에 따라 분리하여 작성함
